pandas/core/array_algos/find_min_subset_sum.py
def find_min_subset_sum(arr, target, closest=False, verbose=False, greedy=False):
    """
    Finds the smallest number of elements whose sum equals the target.

    Parameters:
    arr (list of int): List of integers to find the subset from.
    target (int): The target sum to achieve.
    closest (bool): If True, returns the closest sum if no exact match is found.
    verbose (bool): If True, prints debug information.
    greedy (bool): If True, uses a greedy algorithm instead of the dynamic programming approach.

    Returns:
    dict: A dictionary with the subset, sum, exact_match flag, and number of elements.
    """

    if not isinstance(arr, list) or not all(isinstance(x, int) for x in arr):
        raise ValueError("The input array must be a list of integers.")
    if any(x < 0 for x in arr):
        raise ValueError("Function currently doesn't work with negative integers inside array.")
    if not isinstance(target, int):
        raise ValueError("The target must be an integer.")

    if greedy:
        # Greedy solution
        arr = sorted(arr, reverse=True)
        result = []
        current_sum = 0

        for num in arr:
            if current_sum + num <= target:
                result.append(num)
                current_sum += num
                if verbose:
                    print(f"Adding {num}, current sum: {current_sum}")
                if current_sum == target:
                    return {
                        'subset': result,
                        'sum': current_sum,
                    }

        if closest:
            return {
                'subset': result,
                'sum': current_sum,
            }
        else:
            return {
                'subset': [],
                'sum': 0,
            }
    else:
        # Dynamic programming solution
        dp = [None] * (target + 1)
        dp[0] = []

        for num in arr:
            for i in range(target, num - 1, -1):
                if dp[i - num] is not None and (dp[i] is None or len(dp[i - num]) + 1 < len(dp[i])):
                    dp[i] = dp[i - num] + [num]

        if dp[target] is not None:
            return {
                'subset': dp[target],
                'sum': sum(dp[target]),
            }

        if closest:
            for i in range(target - 1, -1, -1):
                if dp[i] is not None:
                    return {
                        'subset': dp[i],
                        'sum': i,
                    }

        return {
            'subset': [],
            'sum': 0,
        }
